scripts/requirements.txt
atlassian-python-api==3.14.1
certifi==2021.10.8
charset-normalizer==2.0.7
Deprecated==1.2.13
idna==3.3
json2html==1.3.0
oauthlib==3.1.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.7
wrapt==1.13.2